pyproject.toml
[project]
name = "rainbot"
version = "0.1.0"
description = "When it rains, it pours"
readme = "README.md"
authors = [{ name = "Clément Walter", email = "clement0walter@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
  "2captcha-python>=1.5.1",
  "aiohttp>=3.11.16",
  "apscheduler>=3.11.0",
  "beautifulsoup4>=4.13.4",
  "gspread>=6.2.0",
  "html5lib>=1.1",
  "inflection>=0.5.1",
  "numpy>=2.2.4",
  "oauth2client>=4.1.3",
  "pandas>=2.2.3",
  "playwright>=1.49.0",
  "playwright-stealth>=1.0.6",
  "pytz>=2024.1",
  "pytest-asyncio>=1.3.0",
  "python-dotenv>=1.1.0",
  "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = ["ipykernel>=6.29.5", "pytest>=8.3.5"]

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git
    | \.pytest_cache
    | \.venv
    | build
    | dist
    | libs
  )/
)
'''

[tool.pytest.ini_options]
markers = [
  "live: marks tests that run against the live Paris Tennis website (deselect with '-m \"not live\"')",
]
filterwarnings = ["ignore::DeprecationWarning"]
log_cli = true
log_cli_level = "DEBUG"

[tool.isort]
profile = "black"
atomic = true
include_trailing_comma = true
use_parentheses = true
